Rename "interactive" to "combine counts", use concat instead of join
the step no longer requires human intervention, so it's no longer "interactive"
The real purpose is to combine all the read counts from each replicate into a
single file.
